+++ Config/DefaultEngine.ini
@@ -170,6 +170,8 @@ ManualIPAddress=
 +ClassRedirects=(OldName="/Script/project_ND.BTTask_SetTarget",NewName="/Script/project_ND.NDBTTask_SetTarget")
 +ClassRedirects=(OldName="/Script/project_ND.NDBTService_UpdateDistanceFromTarget",NewName="/Script/project_ND.NDBTService_TargetLocationToDestination")
 +ClassRedirects=(OldName="/Script/project_ND.UBTTask_StopAI",NewName="/Script/project_ND.NDBTTask_StopAI")
++PropertyRedirects=(OldName="/Script/project_ND.ItemBaseData.bIsEmpty",NewName="/Script/project_ND.ItemBaseData.bNotUse")
++PropertyRedirects=(OldName="/Script/project_ND.ItemBaseData.bUse",NewName="/Script/project_ND.ItemBaseData.bNotUse")
 
 [/Script/UdpMessaging.UdpMessagingSettings]
 EnabledByDefault=False
