+++ Source/project_ND/PickUpObject/Items/NDPickUpObject_ItemBase.h
@@ -7,38 +7,6 @@
 #include "project_ND/PickUpObject/NDPickUpObject.h"
 #include "NDPickUpObject_ItemBase.generated.h"
 
-UENUM()
-enum EMyEnum
-{
-	HealthPotion,
-	Food,
-	Throwable
-};
-
-USTRUCT(BlueprintType)
-struct FItemBaseData : public FTableRowBase
-{
-	GENERATED_BODY()
-
-	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category="Item", meta=(AllowPrivateAccess))
-	FString ItemName;
-	
-	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category="Item", meta=(AllowPrivateAccess))
-	FString ItemType;
-	
-	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category="Item", meta=(AllowPrivateAccess))
-	float RecoveryAmount;
-	
-	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category="Item", meta=(AllowPrivateAccess))
-	UStaticMesh* ItemMesh;
-
-	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category="Item", meta=(AllowPrivateAccess))
-	TSoftObjectPtr<UTexture2D> ItemThumbnail;
-
-	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category="Item", meta=(AllowPrivateAccess))
-	bool bIsEmpty;
-	
-};
 
 UCLASS()
 class PROJECT_ND_API ANDPickUpObject_ItemBase : public ANDPickUpObject
@@ -53,19 +21,6 @@ protected:
 
 public:
 	virtual void Tick(float DeltaTime) override;
-
-protected:
-	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category="Item", meta=(AllowPrivateAccess))
-	FString ItemType;
-
-	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category="Item", meta=(AllowPrivateAccess))
-	float RecoveryAmount;
-
-	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category="Item", meta=(AllowPrivateAccess))
-	bool bIsEmpty;
-	
-public:
-	virtual void InitializeItem(const struct FItemBaseData& ItemData);
 	
 	virtual void Use();
