+++ Source/project_ND/PickUpObject/Items/NDPickUpObject_ItemBase_HealthPotion.cpp
@@ -20,21 +20,6 @@ void ANDPickUpObject_ItemBase_HealthPotion::Tick(float DeltaTime)
 	Super::Tick(DeltaTime);
 }
 
-void ANDPickUpObject_ItemBase_HealthPotion::InitializeItem(const FItemBaseData& ItemData)
-{
-	Super::InitializeItem(ItemData);
-
-	SetItemName(ItemData.ItemName);
-	ItemType = ItemData.ItemType;
-	RecoveryAmount = ItemData.RecoveryAmount;
-	bIsEmpty = ItemData.bIsEmpty;
-	
-	if (ItemData.ItemMesh)
-	{
-		SetItemMesh(ItemData.ItemMesh);
-	}
-}
-
 void ANDPickUpObject_ItemBase_HealthPotion::Use()
 {
 	Super::Use();
