+++ Source/project_ND/PickUpObject/Items/NDPickUpObject_ItemBase_Throwable.cpp
@@ -55,21 +55,6 @@ void ANDPickUpObject_ItemBase_Throwable::Tick(float DeltaTime)
 	Super::Tick(DeltaTime);
 }
 
-void ANDPickUpObject_ItemBase_Throwable::InitializeItem(const FItemBaseData& ItemData)
-{
-	Super::InitializeItem(ItemData);
-
-	SetItemName(ItemData.ItemName);
-	ItemType = ItemData.ItemType;
-	RecoveryAmount = ItemData.RecoveryAmount;
-	bIsEmpty = ItemData.bIsEmpty;
-	
-	if (ItemData.ItemMesh)
-	{
-		SetItemMesh(ItemData.ItemMesh);
-	}
-}
-
 void ANDPickUpObject_ItemBase_Throwable::Use()
 {
 	Super::Use();
