+++ Source/project_ND/PickUpObject/Items/NDPickUpObject_ItemBase_Throwable.h
@@ -32,8 +32,6 @@ public:
 
 	virtual void Tick(float DeltaTime) override;
 
-	virtual void InitializeItem(const struct FItemBaseData& ItemData);
-
 	virtual void Use() override;
 
 	bool bCanNoise = true;
